Samples go in the reservoir
